+++ Assets/Scripts/Meta/InventorySystem.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 524256769e9f1904d9d7965efee92142
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
