+++ Assets/Scripts/Meta/InventorySystem/CardSO.cs
@@ -1,0 +1,19 @@
+using Meta.Interfaces;
+using UnityEngine;
+
+namespace Meta.InventorySystem
+{
+    [CreateAssetMenu(fileName = "NewCard", menuName = "CardSystem/Card")]
+    public class CardSO : ScriptableObject, ICard
+    {
+        [SerializeField] string cardTitle;
+        [SerializeField] int level;
+        [SerializeField] int apCost;
+        [SerializeField] GameObject unitPrefab;
+
+        public string Name => cardTitle;
+        public int CardLevel => level;
+        public int ApCost => apCost;
+        public GameObject CardObject => unitPrefab;
+    }
+}
